backend/requirements-dev.txt
-r requirements.txt
# Linter + formatter (single binary, replaces flake8 + black + isort).
ruff==0.15.12
# Static type-checker with Django support. See pyproject.toml.
mypy==1.20.2
django-stubs==5.1.3
djangorestframework-stubs==3.15.2
Django>=5.2.4,<6.0.0
djangorestframework>=3.16.0,<4.0.0
python-dotenv>=1.1.1,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
djangorestframework-simplejwt[crypto]==5.5.0
Pillow>=10.0.0,<11.0.0
python-json-logger>=2.0.7,<3.0.0
django-filter==25.1
django-debug-toolbar==5.2.0
django-cors-headers>=4.4.0,<5.0.0
django-model-utils>=4.3.1,<5.0.0
django-model-utils>=4.3.1,<5.0.0
